flaskapp-old/requirements.txt
flask==3.0.0
python-dotenv==1.0.0
pillow==10.2.0
pypdf2==3.0.1
langchain==0.1.0
langchain-openai==0.0.2
faiss-cpu==1.7.4
openai==1.10.0